feat: Supports annotations on assignments & object properties
The generated code is not *great*, it creates a lot of new bindings and
is overall kinda hard to read. A helper function would help, but there's
no documentation on how you'd go about adding those with Babel.

Currently properties might be visited twice, in some obscure cases. The
result is still the same (transformations are idempotent), but the
generated code becomes even worse.