(build) Makes it simpler to build all projects
Each project now includes its own Makefiles, which define rules about how to build, lint, clean, etc. that project. The root Makefile then runs all of the tasks defined in the subproject Makefile for the common commands.

Packages that get run are controlled by the MM_PACKAGES environment variable. By default, all of the stable packages (in packages/) get ran when running a command, unless MM_PACKAGES is provided in the environment.

If the MM_EXPERIMENTAL_PACKAGES environment variable is "true", the packages are initialised to all projects in the repository, instead, including experimental ones.