(new static) Adds a concept of "class" that shows prototype properties
JS, implementation-wise, is a prototype-based language, so each object defines its own layout. But people think about programs, many times, in terms of classes. It makes sense to show the layout of a function object and the layout of its prototype property together when this function matches this concept of class, so it's closer to how people think about it.